Imp. Deletion Bug Fix & Concurrent Deletion Bug Fix
There were some bugs in the file and metadata file deletion process that
in some occasions could lead to undeletable files. This has been fixed.
In addition, I fixed another glaring issue when it came to the deletion
process; I made it impossible to delete a file that is either printing
or still in the queue. At some point I will need to customize the
messages, because now there is only one standard failure message, from
which it is impossible to determine the cause, especially for the
typical user.